fix: address cdp url-state review
