+++ ui-dashboard/src/app/cdps/_components/cdps-page-client.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useMemo } from "react";
+import { Suspense, useMemo } from "react";
 import { useNetwork } from "@/components/network-provider";
 import { EmptyBox, ErrorBox, Skeleton } from "@/components/feedback";
 import { useGQL } from "@/lib/graphql";
@@ -200,10 +200,23 @@ export function CdpsPageClient() {
           />
         ))}
       </div>
-      <CdpAllTransactionsTable
-        collaterals={collaterals}
-        chainId={network.chainId}
-      />
+      <Suspense fallback={<CdpTransactionsTableFallback />}>
+        <CdpAllTransactionsTable
+          collaterals={collaterals}
+          chainId={network.chainId}
+        />
+      </Suspense>
     </div>
+  );
+}
+
+function CdpTransactionsTableFallback() {
+  return (
+    <section>
+      <h2 className="text-lg font-semibold text-white mb-3">
+        Recent CDP Transactions
+      </h2>
+      <Skeleton rows={6} />
+    </section>
   );
 }
